+++ mayan/apps/checkouts/apps.py
@@ -11,10 +11,11 @@ from django.utils.translation import ugettext_lazy as _
 
 from acls import ModelPermission
 from common import MayanAppConfig, menu_facet, menu_main, menu_sidebar
-from common.classes import DashboardWidget
+from common.dashboards import dashboard_main
 from mayan.celery import app
 from rest_api.classes import APIEndPoint
 
+from .dashboard_widgets import widget_checkouts
 from .handlers import check_new_version_creation
 from .links import (
     link_checkin_document, link_checkout_document, link_checkout_info,
@@ -48,13 +49,6 @@ class CheckoutsApp(MayanAppConfig):
         )
 
         DocumentCheckout = self.get_model('DocumentCheckout')
-
-        DashboardWidget(
-            icon='fa fa-shopping-cart',
-            queryset=DocumentCheckout.objects.all(),
-            label=_('Checkedout documents'),
-            link=reverse_lazy('checkouts:checkout_list')
-        )
 
         Document.add_to_class(
             'check_in',
@@ -113,6 +107,8 @@ class CheckoutsApp(MayanAppConfig):
                 },
             }
         )
+
+        dashboard_main.add_widget(order=-1, widget=widget_checkouts)
 
         menu_facet.bind_links(links=(link_checkout_info,), sources=(Document,))
         menu_main.bind_links(links=(link_checkout_list,), position=98)